Sushma|Adding verifications for observations

+++ spec/spec_helper.rb
@@ -15,16 +15,17 @@ Capybara.default_wait_time = 10
 Capybara.ignore_hidden_elements = false
 Capybara.save_and_open_page_path = File.expand_path("../../screenshots", __FILE__)
 
-headless = Headless.new
+headless = Headless.new if Settings.headless
+
 Debugger.start
 
 RSpec.configure do |config|
     config.include AppsAwareness
     config.before(:all) do
         Selenium::WebDriver::Firefox::Binary.path=Settings.firefox_path
-        headless.start
+        headless.start if Settings.headless
     end
     config.after(:all) do
-        headless.destroy
+        headless.destroy if Settings.headless
     end
 end
